common, wsd: clean up redundant casts

+++ wsd/Admin.cpp
@@ -342,7 +342,7 @@ Admin::Admin() :
     _totalSysMem = Util::getTotalSystemMemory();
     LOG_TRC("Total system memory : " << _totalSysMem);
 
-    const auto memLimit = LOOLWSD::getConfigValue<double>("memproportion", static_cast<double>(0.0));
+    const auto memLimit = LOOLWSD::getConfigValue<double>("memproportion", 0.0);
     _totalAvailMem = _totalSysMem;
     if (memLimit != 0.0)
         _totalAvailMem = _totalSysMem * memLimit/100.;
@@ -554,7 +554,7 @@ void Admin::triggerMemoryCleanup(size_t totalMem)
 {
     LOG_TRC("Total memory we are consuming (in kB): " << totalMem);
     // Trigger mem cleanup when we are consuming too much memory (as configured by sysadmin)
-    const auto memLimit = LOOLWSD::getConfigValue<double>("memproportion", static_cast<double>(0.0));
+    const auto memLimit = LOOLWSD::getConfigValue<double>("memproportion", 0.0);
     LOG_TRC("Mem proportion for LOOL configured : " << memLimit);
     if (memLimit == 0.0 || _totalSysMem == 0)
     {
